move the projection endpoint into the api layer
Whitelisted endpoints belong in billing/api; the projection package stays
domain-only with no HTTP surface. The boundary test caught this.
