+++ central/billing/api/admin/__init__.py
@@ -22,6 +22,7 @@ from central.billing.api.admin.catalog import (
 	get_trial_detail,
 	update_plan_rate,
 )
+from central.billing.api.admin.projection import project_team
 from central.billing.api.admin.revenue import (
 	get_cluster_breakdown,
 	get_free_trial_costs,
@@ -43,6 +44,7 @@ from central.billing.api.admin.teams import (
 )
 
 __all__ = [
+	"project_team",
 	"adjust_team_credits",
 	"create_configured_plan",
 	"get_catalog",
